fix: align JS implementation with Python via cross-language parity harness
Add tests/parity/ — a pytest harness that runs the browser JS (docs/js)
under Node on Python-generated inputs and compares outputs against the
Python reference implementations, including an end-to-end run through
the real maia3-5m.onnx model.

Misalignments found and fixed:

- pgn.js: clock annotations ([%clk]/[%clk_opp]) were never parsed (the
  whitespace tokenizer could not match brace comments), so clocks always
  defaulted to 300/0. Now uses chess.js getComments() keyed by FEN,
  matching python-chess node-comment semantics.
- pgn.js/tensor.js: replay boards ignored [FEN] headers; positions now
  carry fenBefore from the verbose history and batch builders start the
  replay from it. Invalid PGNs return null instead of throwing,
  mirroring python-chess.
- scoring.js: computeScoreStreaming stored results in a Float32Array
  while Python returns float64 (~1e-8 relative error); now Float64Array.
- tensor.js: buildBatchTensorSingleColor token stacking used an
  out-of-range subarray on per-position tokens, leaving all sampled
  positions after the first with all-zero tokens in the per-color ELO
  refinement path. Fixed to stack-then-tile like the 1D builder.

+++ docs/js/scoring.js
@@ -9,7 +9,8 @@ const FIDELITY = 50;
 // Compute blended top-1 accuracy + MRR score
 // Processes one ELO at a time to save memory
 function computeScoreStreaming(logitsAllElos, humanMoves, legalMasks, nPos, nElo) {
-  const scores = new Float32Array(nElo);
+  // Float64 to match numpy float64 output of batch_inference.py:_compute_score
+  const scores = new Float64Array(nElo);
   if (nPos === 0) return scores;
 
   for (let e = 0; e < nElo; e++) {
